+++ test/compute-download-urls-test.js
@@ -1,6 +1,6 @@
-var assert = require('assert');
+const assert = require('assert');
 
-var computeDownloadUrls;
+let computeDownloadUrls;
 
 /**
  * Tests for the `computeDownloadUrls` module.
@@ -9,253 +9,253 @@ var computeDownloadUrls;
  * the desired binary files. Just that the logic contained in the module, specifically for
  * handling when paths formats differ between versions of the same driver.
  */
-describe('compute-download-urls', function() {
+describe('compute-download-urls', () => {
   // Allow tests to mock `process.platform`
-  before(function() {
+  before(function () {
     this.originalPlatform = Object.getOwnPropertyDescriptor(process, 'platform');
   });
-  after(function() {
+  after(function () {
     Object.defineProperty(process, 'platform', this.originalPlatform);
   });
 
   // Ensure that any internal state of the module is clean for each test
-  beforeEach(function() {
+  beforeEach(() => {
     computeDownloadUrls = require('../lib/compute-download-urls');
   });
-  afterEach(function() {
+  afterEach(() => {
     delete require.cache[require.resolve('../lib/compute-download-urls')];
   });
 
-  var opts;
+  let opts;
 
-  describe('selenium-jar', function() {
-    it('basic version', function() {
-      var actual = computeDownloadUrls({
+  describe('selenium-jar', () => {
+    it('basic version', async () => {
+      const actual = await computeDownloadUrls({
         seleniumVersion: '1.0',
         seleniumBaseURL: 'https://localhost',
-        drivers: {}
+        drivers: {},
       });
 
-      assert.equal(actual.selenium, 'https://localhost/1.0/selenium-server-standalone-1.0.jar');
+      assert.strictEqual(actual.selenium, 'https://localhost/1.0/selenium-server-standalone-1.0.jar');
     });
 
-    it('version with patch', function() {
-      var actual = computeDownloadUrls({
+    it('version with patch', async () => {
+      const actual = await computeDownloadUrls({
         seleniumVersion: '1.0.1',
         seleniumBaseURL: 'https://localhost',
-        drivers: {}
+        drivers: {},
       });
 
-      assert.equal(actual.selenium, 'https://localhost/1.0/selenium-server-standalone-1.0.1.jar');
+      assert.strictEqual(actual.selenium, 'https://localhost/1.0/selenium-server-standalone-1.0.1.jar');
     });
 
-    it('version with beta string', function() {
-      var actual = computeDownloadUrls({
+    it('version with beta string', async () => {
+      const actual = await computeDownloadUrls({
         seleniumVersion: '3.0.0-beta2',
         seleniumBaseURL: 'https://localhost',
-        drivers: {}
+        drivers: {},
       });
 
-      assert.equal(actual.selenium, 'https://localhost/3.0-beta2/selenium-server-standalone-3.0.0-beta2.jar');
+      assert.strictEqual(actual.selenium, 'https://localhost/3.0-beta2/selenium-server-standalone-3.0.0-beta2.jar');
     });
   });
 
-  describe('chrome', function() {
-    beforeEach(function() {
+  describe('chrome', () => {
+    beforeEach(() => {
       opts = {
         seleniumVersion: '1.0',
         seleniumBaseURL: 'https://localhost',
         drivers: {
-          chrome: {}
-        }
+          chrome: {},
+        },
       };
     });
 
-    describe('linux', function() {
-      before(function() {
+    describe('linux', () => {
+      before(() => {
         Object.defineProperty(process, 'platform', {
-          value: 'linux'
+          value: 'linux',
         });
       });
 
-      it('x32 for versions < 2.34', function() {
+      it('x32 for versions < 2.34', async () => {
         opts.drivers.chrome = {
           baseURL: 'https://localhost',
           version: '2.0',
-          arch: 'x32'
-        }
+          arch: 'x32',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.chrome, 'https://localhost/2.0/chromedriver_linux32.zip');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.chrome, 'https://localhost/2.0/chromedriver_linux32.zip');
       });
 
-      it('x64', function() {
+      it('x64', async () => {
         opts.drivers.chrome = {
           baseURL: 'https://localhost',
           version: '2.0',
-          arch: 'x64'
-        }
+          arch: 'x64',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.chrome, 'https://localhost/2.0/chromedriver_linux64.zip');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.chrome, 'https://localhost/2.0/chromedriver_linux64.zip');
       });
     });
 
-    describe('mac', function() {
-      before(function() {
+    describe('mac', () => {
+      before(() => {
         Object.defineProperty(process, 'platform', {
-          value: 'darwin'
+          value: 'darwin',
         });
       });
 
-      it('Use `mac32` for versions < 2.23', function() {
+      it('Use `mac32` for versions < 2.23', async () => {
         opts.drivers.chrome = {
           baseURL: 'https://localhost',
           version: '2.22',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.chrome, 'https://localhost/2.22/chromedriver_mac32.zip');
-      })
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.chrome, 'https://localhost/2.22/chromedriver_mac32.zip');
+      });
 
-      it('Use `mac64` for versions >= 2.23', function() {
+      it('Use `mac64` for versions >= 2.23', async () => {
         opts.drivers.chrome = {
           baseURL: 'https://localhost',
           version: '2.23',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.chrome, 'https://localhost/2.23/chromedriver_mac64.zip');
-      })
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.chrome, 'https://localhost/2.23/chromedriver_mac64.zip');
+      });
     });
 
-    describe('win', function() {
-      before(function() {
+    describe('win', () => {
+      before(() => {
         Object.defineProperty(process, 'platform', {
-          value: 'win32'
+          value: 'win32',
         });
       });
 
-      it('basic version', function() {
+      it('basic version', async () => {
         opts.drivers.chrome = {
           baseURL: 'https://localhost',
           version: '2.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.chrome, 'https://localhost/2.0/chromedriver_win32.zip');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.chrome, 'https://localhost/2.0/chromedriver_win32.zip');
       });
     });
   });
 
-  describe('firefox', function() {
-    beforeEach(function() {
+  describe('firefox', () => {
+    beforeEach(() => {
       opts = {
         seleniumVersion: '1.0',
         seleniumBaseURL: 'https://localhost',
         drivers: {
-          firefox: {}
-        }
+          firefox: {},
+        },
       };
     });
 
-    describe('linux', function() {
-      before(function() {
+    describe('linux', () => {
+      before(() => {
         Object.defineProperty(process, 'platform', {
-          value: 'linux'
+          value: 'linux',
         });
       });
 
-      it('uses `wires` name for versions < 0.8.0', function() {
+      it('uses `wires` name for versions < 0.8.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.7.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.firefox, 'https://localhost/v0.7.0/wires-0.7.0-linux64.gz');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.firefox, 'https://localhost/v0.7.0/wires-0.7.0-linux64.gz');
       });
 
-      it('uses `geckodriver` name for versions >= 0.8.0', function() {
+      it('uses `geckodriver` name for versions >= 0.8.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.8.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.firefox, 'https://localhost/v0.8.0/geckodriver-0.8.0-linux64.gz');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.firefox, 'https://localhost/v0.8.0/geckodriver-0.8.0-linux64.gz');
       });
 
-      it('uses correct directory for 0.3.0', function() {
+      it('uses correct directory for 0.3.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.3.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.firefox, 'https://localhost/0.3.0/wires-0.3.0-linux64.gz');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.firefox, 'https://localhost/0.3.0/wires-0.3.0-linux64.gz');
       });
 
-      it('uses leading `v` in version string when >= 0.9.0', function() {
+      it('uses leading `v` in version string when >= 0.9.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.9.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.firefox, 'https://localhost/v0.9.0/geckodriver-v0.9.0-linux64.tar.gz');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.firefox, 'https://localhost/v0.9.0/geckodriver-v0.9.0-linux64.tar.gz');
       });
 
-      it('uses plain version string when < 0.9.0', function() {
+      it('uses plain version string when < 0.9.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.7.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.firefox, 'https://localhost/v0.7.0/wires-0.7.0-linux64.gz');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.firefox, 'https://localhost/v0.7.0/wires-0.7.0-linux64.gz');
       });
 
-      it('uses `.gz` file extension for versions < 0.9.0', function() {
+      it('uses `.gz` file extension for versions < 0.9.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.8.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('.gz') > 0);
         assert(actual.firefox.indexOf('.tar.gz') === -1);
       });
 
-      it('uses `.tar.gz` file extension for versions >= 0.9.0', function() {
+      it('uses `.tar.gz` file extension for versions >= 0.9.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.9.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('.tar.gz') > 0);
       });
 
-      it('throws if asking a < 0.11.0 version and an arch which is not x64', function() {
+      it('throws if asking a < 0.11.0 version and an arch which is not x64', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.9.0',
-          arch: 'x86'
+          arch: 'x86',
         };
 
         try {
-          computeDownloadUrls(opts);
+          await computeDownloadUrls(opts);
           throw new Error('Error not thrown');
         } catch (err) {
           if (err && err.message === 'Only x64 architecture is available for Firefox < 0.11.0') {
@@ -265,153 +265,153 @@ describe('compute-download-urls', function() {
         }
       });
 
-      it('gets the right arch when arch is x86', function() {
+      it('gets the right arch when arch is x86', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.11.0',
-          arch: 'x86'
+          arch: 'x86',
         };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('-linux32.tar.gz') > 0);
       });
 
-      it('gets the right arch when arch is x64', function() {
+      it('gets the right arch when arch is x64', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.11.0',
-          arch: 'x64'
+          arch: 'x64',
         };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('-linux64.tar.gz') > 0);
       });
 
-      it('gets the right arch when arch is x32', function() {
+      it('gets the right arch when arch is x32', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.11.0',
-          arch: 'x32'
+          arch: 'x32',
         };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('-linux32.tar.gz') > 0);
       });
     });
 
-    describe('mac', function() {
-      before(function() {
+    describe('mac', () => {
+      before(() => {
         Object.defineProperty(process, 'platform', {
-          value: 'darwin'
+          value: 'darwin',
         });
       });
 
-      it('uses `OSX` platform for versions < 0.9.0', function() {
+      it('uses `OSX` platform for versions < 0.9.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.8.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('OSX') > 0);
       });
 
-      it('uses `mac` platform for versions == 0.9.0', function() {
+      it('uses `mac` platform for versions == 0.9.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.9.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('mac') > 0);
         assert(actual.firefox.indexOf('macos') === -1);
       });
 
-      it('uses `macos` platform for versions >= 0.10.0', function() {
+      it('uses `macos` platform for versions >= 0.10.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.10.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('macos') > 0);
       });
 
-      it('uses `osx` platform for versions <= 0.6.0', function() {
+      it('uses `osx` platform for versions <= 0.6.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.5.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('osx') > 0);
       });
     });
 
-    describe('win', function() {
-      before(function() {
+    describe('win', () => {
+      before(() => {
         Object.defineProperty(process, 'platform', {
-          value: 'win32'
+          value: 'win32',
         });
       });
 
-      it('uses leading `v` in version string when == 0.5.0', function() {
+      it('uses leading `v` in version string when == 0.5.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.5.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.firefox, 'https://localhost/v0.5.0/wires-v0.5.0-win.zip');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.firefox, 'https://localhost/v0.5.0/wires-v0.5.0-win.zip');
       });
 
-      it('gets the right arch when arch is x32 and version >= 0.11.0', function() {
+      it('gets the right arch when arch is x32 and version >= 0.11.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.11.1',
-          arch: 'x32'
+          arch: 'x32',
         };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('win32.zip') >= 0);
       });
 
-      it('gets the right arch when arch is x64 and version >= 0.11.0', function() {
+      it('gets the right arch when arch is x64 and version >= 0.11.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.11.0',
-          arch: 'x64'
+          arch: 'x64',
         };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('win64.zip') >= 0);
       });
 
-      it('gets the 32 bits version when no arch specified version >= 0.11.0', function() {
+      it('gets the 32 bits version when no arch specified version >= 0.11.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.11.0',
-          arch: ''
+          arch: '',
         };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('win32.zip') >= 0);
       });
 
-      it('throws if asking the 32bit version for 0.9.0/0.10.0', function() {
+      it('throws if asking the 32bit version for 0.9.0/0.10.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.10.0',
-          arch: 'x32'
+          arch: 'x32',
         };
 
         try {
-          computeDownloadUrls(opts);
+          await computeDownloadUrls(opts);
           throw new Error('Error not thrown');
         } catch (err) {
           if (err && err.message === 'Only x64 architecture is available for Firefox 0.9.0 and 0.10.0') {
@@ -421,15 +421,15 @@ describe('compute-download-urls', function() {
         }
       });
 
-      it('throws if asking the 64bit version for < 0.9.0', function() {
+      it('throws if asking the 64bit version for < 0.9.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.8.0',
-          arch: 'x64'
+          arch: 'x64',
         };
 
         try {
-          computeDownloadUrls(opts);
+          await computeDownloadUrls(opts);
           throw new Error('Error not thrown');
         } catch (err) {
           if (err && err.message === 'Only 32 bits architectures are available for Firefox <= 0.8.0') {
@@ -439,237 +439,233 @@ describe('compute-download-urls', function() {
         }
       });
 
-      it('uses `win32` name for versions 0.8.0 & 0.7.1', function() {
+      it('uses `win32` name for versions 0.8.0 & 0.7.1', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.7.1',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('-win32.zip') > 0);
       });
 
-      it('uses `windows` name for version 0.3.0', function() {
+      it('uses `windows` name for version 0.3.0', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.3.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('-windows.zip') > 0);
       });
 
-      it('uses `win` name for versions other versions', function() {
+      it('uses `win` name for versions other versions', async () => {
         opts.drivers.firefox = {
           baseURL: 'https://localhost',
           version: '0.5.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.firefox.indexOf('-win.zip') > 0);
       });
-
     });
   });
 
-  describe('ie', function() {
-    before(function() {
+  describe('ie', () => {
+    before(() => {
       Object.defineProperty(process, 'platform', {
-        value: 'win32'
+        value: 'win32',
       });
     });
 
-    beforeEach(function() {
+    beforeEach(() => {
       opts = {
         seleniumVersion: '1.0',
         seleniumBaseURL: 'https://localhost',
         drivers: {
-          ie: {}
-        }
+          ie: {},
+        },
       };
     });
 
-    it('uses `Win32` platform when arch == ia32', function() {
+    it('uses `Win32` platform when arch == ia32', async () => {
       opts.drivers.ie = {
         baseURL: 'https://localhost',
         version: '2.20.0',
-        arch: 'ia32'
-      }
+        arch: 'ia32',
+      };
 
-      var actual = computeDownloadUrls(opts);
+      const actual = await computeDownloadUrls(opts);
       assert(actual.ie.indexOf('IEDriverServer_Win32') > 0);
     });
 
-    it('uses `x64` platform when arch == x64', function() {
+    it('uses `x64` platform when arch == x64', async () => {
       opts.drivers.ie = {
         baseURL: 'https://localhost',
         version: '2.20.0',
-        arch: 'x64'
-      }
+        arch: 'x64',
+      };
 
-      var actual = computeDownloadUrls(opts);
+      const actual = await computeDownloadUrls(opts);
       assert(actual.ie.indexOf('IEDriverServer_x64') > 0);
     });
 
-    it('uses `major.minor` folder for `major.minor.patch` version', function() {
+    it('uses `major.minor` folder for `major.minor.patch` version', async () => {
       opts.drivers.ie = {
         baseURL: 'https://localhost',
         version: '2.20.1',
-        arch: 'x64'
-      }
+        arch: 'x64',
+      };
 
-      var actual = computeDownloadUrls(opts);
+      const actual = await computeDownloadUrls(opts);
       assert(actual.ie.indexOf('/2.20/') > 0);
       assert(actual.ie.indexOf('2.20.1.zip') > 0);
     });
   });
 
-  describe('edge', function() {
-    before(function() {
+  describe('edge', () => {
+    before(() => {
       Object.defineProperty(process, 'platform', {
-        value: 'win32'
+        value: 'win32',
       });
-
     });
 
-    beforeEach(function() {
+    beforeEach(() => {
       opts = {
         seleniumVersion: '1.0',
         seleniumBaseURL: 'https://localhost',
         drivers: {
-          edge: {}
-        }
+          edge: {},
+        },
       };
     });
 
-    var releases = require('../lib/microsoft-edge-releases');
+    const releases = require('../lib/microsoft-edge-releases');
 
-    Object.keys(releases).forEach(function (version) {
-      it('uses version `' + version + '` correct url', function() {
+    Object.keys(releases).forEach((version) => {
+      it('uses version `' + version + '` correct url', async () => {
         opts.drivers.edge = { version: version };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.edge, releases[version].url);
-      })
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.edge, releases[version].url);
+      });
     });
 
-    Object.keys(releases).forEach(function (version) {
-      it('uses version `' + version + '` correct extension', function() {
+    Object.keys(releases).forEach((version) => {
+      it('uses version `' + version + '` correct extension', async () => {
         opts.drivers.edge = { version: version };
 
-        var actual = computeDownloadUrls(opts);
+        const actual = await computeDownloadUrls(opts);
         assert(actual.edge.indexOf(releases[version].extension) > 0);
-      })
+      });
     });
 
-    it('throws for unknown releases', function() {
-      ['1.0', '2.3', '10'].forEach(function (version) {
-        opts.drivers.edge = { version: version };
+    it('throws for unknown releases', (done) => {
+      opts.drivers.edge = { version: '10' };
 
-        assert.throws(function () { computeDownloadUrls(opts) });
-      })
+      computeDownloadUrls(opts).catch(() => done());
     });
   });
 
-  describe('chromiumedge', function() {
-    beforeEach(function() {
+  describe('chromiumedge', () => {
+    beforeEach(() => {
       opts = {
         seleniumVersion: '1.0',
         seleniumBaseURL: 'https://localhost',
         drivers: {
-          chromiumedge: {}
-        }
+          chromiumedge: {},
+        },
       };
     });
 
-    describe('linux', function() {
-      before(function() {
+    describe('linux', () => {
+      before(() => {
         Object.defineProperty(process, 'platform', {
-          value: 'linux'
+          value: 'linux',
         });
       });
 
-      it('throws for x32 arch', function() {
+      it('throws for x32 arch', (done) => {
         opts.drivers.chromiumedge = {
           baseURL: 'https://localhost',
           version: '86.0.600.0',
-          arch: 'x32'
-        }
+          arch: 'x32',
+        };
 
-        assert.throws(() => computeDownloadUrls(opts), 'Only x64 is supported for linux');
+        computeDownloadUrls(opts).catch(() => done());
       });
 
-      it('x64', function() {
+      it('x64', async () => {
         opts.drivers.chromiumedge = {
           baseURL: 'https://localhost',
           version: '86.0.600.0',
-          arch: 'x64'
-        }
+          arch: 'x64',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.chromiumedge, 'https://localhost/86.0.600.0/edgedriver_linux64.zip');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.chromiumedge, 'https://localhost/86.0.600.0/edgedriver_linux64.zip');
       });
     });
 
-    describe('mac', function() {
-      before(function() {
+    describe('mac', () => {
+      before(() => {
         Object.defineProperty(process, 'platform', {
-          value: 'darwin'
+          value: 'darwin',
         });
       });
 
-      it('throws for x32 arch', function() {
+      it('throws for x32 arch', (done) => {
         opts.drivers.chromiumedge = {
           baseURL: 'https://localhost',
           version: '86.0.600.0',
-          arch: 'x32'
-        }
+          arch: 'x32',
+        };
 
-        assert.throws(() => computeDownloadUrls(opts), 'Only x64 is supported for mac');
-      })
+        computeDownloadUrls(opts).catch(() => done());
+      });
 
-      it('x64', function() {
+      it('x64', async () => {
         opts.drivers.chromiumedge = {
           baseURL: 'https://localhost',
           version: '86.0.600.0',
-          arch: 'x64'
-        }
+          arch: 'x64',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.chromiumedge, 'https://localhost/86.0.600.0/edgedriver_mac64.zip');
-      })
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.chromiumedge, 'https://localhost/86.0.600.0/edgedriver_mac64.zip');
+      });
     });
 
-    describe('win', function() {
-      before(function() {
+    describe('win', () => {
+      before(() => {
         Object.defineProperty(process, 'platform', {
-          value: 'win32'
+          value: 'win32',
         });
       });
 
-      it('x32', function() {
+      it('x32', async () => {
         opts.drivers.chromiumedge = {
           baseURL: 'https://localhost',
           version: '86.0.600.0',
-          arch: 'x32'
-        }
+          arch: 'x32',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.chromiumedge, 'https://localhost/86.0.600.0/edgedriver_win32.zip');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.chromiumedge, 'https://localhost/86.0.600.0/edgedriver_win32.zip');
       });
 
-      it('x64', function() {
+      it('x64', async () => {
         opts.drivers.chromiumedge = {
           baseURL: 'https://localhost',
           version: '86.0.600.0',
-          arch: ''
-        }
+          arch: '',
+        };
 
-        var actual = computeDownloadUrls(opts);
-        assert.equal(actual.chromiumedge, 'https://localhost/86.0.600.0/edgedriver_win64.zip');
+        const actual = await computeDownloadUrls(opts);
+        assert.strictEqual(actual.chromiumedge, 'https://localhost/86.0.600.0/edgedriver_win64.zip');
       });
     });
   });